refactor(cli): break cmd_support/collectors cycle via support package
Move CollectorResult to support/__init__.py so collectors.py imports
from its own package instead of from cmd_support.
