+++ observal_cli/cmd_support.py
@@ -21,7 +21,6 @@ import tarfile
 import tempfile
 import time
 from concurrent.futures import ThreadPoolExecutor
-from dataclasses import dataclass
 from datetime import UTC, datetime
 from pathlib import Path
 
@@ -74,31 +73,7 @@ SIZE_BUDGET_BYTES = 100 * 1024 * 1024  # 100 MB uncompressed warning threshold
 # ── CollectorResult ──────────────────────────────────────────────────
 
 
-@dataclass
-class CollectorResult:
-    """Result from a single diagnostic collector."""
-
-    name: str  # e.g. "versions", "health_postgres"
-    ok: bool
-    duration_ms: int
-    data: dict | list | str | None
-    error: str | None = None
-
-    @property
-    def target_path(self) -> str:
-        """Relative path in the archive, e.g. 'versions/app.json'."""
-        # Map collector names to archive paths
-        _path_map: dict[str, str] = {
-            "versions": "versions/app.json",
-            "health": "health/health.json",
-            "config": "config/config.json",
-            "aggregates": "aggregates/aggregates.json",
-            "errors": "errors/recent_errors.json",
-            "logs": "logs/recent.ndjson",
-            "config_allowlisted": "config/config.json",
-            "system_info": "system/system.json",
-        }
-        return _path_map.get(self.name, f"{self.name}.json")
+from observal_cli.support import CollectorResult  # noqa: F401 -- re-exported
 
 
 # ── Local collectors ─────────────────────────────────────────────────
